Refactor clan and material insertion logic to handle conflicts and streamline pending entries deletion

+++ database/create-function/handle_contract_event_testnet.sql
@@ -5,13 +5,23 @@ BEGIN
   -- ClanEmblems
   IF NEW.contract_address = '0x9322C4A5E5725262C9960aDE87259d1cE2812412' THEN
     IF NEW.name = 'ClanCreated' THEN
-      INSERT INTO clans (chain_id, id, owner, name, logo_image_url, logo_thumbnail_url, description, created_at)
-      SELECT NEW.chain_id, NEW.args->>'clanId', NEW.args->>'clanOwner', pc.name, pc.logo_image_url, pc.logo_thumbnail_url, pc.description, NOW()
-      FROM pending_clans pc
-      WHERE pc.metadata_hash = NEW.args->>'metadataHash';
+      IF EXISTS (SELECT 1 FROM pending_clans WHERE metadata_hash = NEW.args->>'metadataHash') THEN
+        INSERT INTO clans (chain_id, id, owner, name, logo_image_url, logo_thumbnail_url, description)
+        SELECT NEW.chain_id, NEW.args->>'clanId', NEW.args->>'clanOwner', pc.name, pc.logo_image_url, pc.logo_thumbnail_url, pc.description
+        FROM pending_clans pc
+        WHERE pc.metadata_hash = NEW.args->>'metadataHash'
+        ON CONFLICT (chain_id, id) DO UPDATE
+        SET owner = NEW.args->>'clanOwner';
 
-      DELETE FROM pending_clans
-      WHERE metadata_hash = NEW.args->>'metadataHash';
+        DELETE FROM pending_clans
+        WHERE metadata_hash = NEW.args->>'metadataHash';
+      ELSE
+        INSERT INTO clans (chain_id, id, owner, name)
+        VALUES (NEW.chain_id, NEW.args->>'clanId', NEW.args->>'clanOwner', 'Unnamed Clan')
+        ON CONFLICT (chain_id, id) DO UPDATE
+        SET owner = NEW.args->>'clanOwner';
+      END IF;
+
     ELSIF NEW.name = 'ClanDeleted' THEN
       DELETE FROM clans
       WHERE chain_id = NEW.chain_id
@@ -22,13 +32,23 @@ BEGIN
   -- MaterialFactory
   IF NEW.contract_address = '0xc78c189C24379857A80635624877E02306de3EE1' THEN
     IF NEW.name = 'MaterialCreated' THEN
-      INSERT INTO materials (chain_id, address, owner, name, symbol, logo_image_url, logo_thumbnail_url, description, created_at)
-      SELECT NEW.chain_id, NEW.args->>'materialAddress', NEW.args->>'materialOwner', pm.name, pm.symbol, pm.logo_image_url, pm.logo_thumbnail_url, pm.description, NOW()
-      FROM pending_materials pm
-      WHERE pm.metadata_hash = NEW.args->>'metadataHash';
+      IF EXISTS (SELECT 1 FROM pending_materials WHERE metadata_hash = NEW.args->>'metadataHash') THEN
+        INSERT INTO materials (chain_id, address, owner, name, symbol, logo_image_url, logo_thumbnail_url, description)
+        SELECT NEW.chain_id, NEW.args->>'materialAddress', NEW.args->>'materialOwner', NEW.args->>'name', NEW.args->>'symbol', pm.logo_image_url, pm.logo_thumbnail_url, pm.description
+        FROM pending_materials pm
+        WHERE pm.metadata_hash = NEW.args->>'metadataHash'
+        ON CONFLICT (chain_id, address) DO UPDATE
+        SET owner = NEW.args->>'materialOwner';
 
-      DELETE FROM pending_materials
-      WHERE metadata_hash = NEW.args->>'metadataHash';
+        DELETE FROM pending_materials
+        WHERE metadata_hash = NEW.args->>'metadataHash';
+      ELSE
+        INSERT INTO materials (chain_id, address, owner, name, symbol)
+        VALUES (NEW.chain_id, NEW.args->>'materialAddress', NEW.args->>'materialOwner', NEW.args->>'name', NEW.args->>'symbol')
+        ON CONFLICT (chain_id, address) DO UPDATE
+        SET owner = NEW.args->>'materialOwner';
+      END IF;
+
     ELSIF NEW.name = 'MaterialDeleted' THEN
       DELETE FROM materials
       WHERE chain_id = NEW.chain_id
